XIVY-16824 implements feedback
