--- packages/database-editor/src/components/DatabaseTableInfo.css
@@ -1,4 +1,0 @@
-.table-info {
-  max-height: 70vh;
-  overflow: scroll;
-}
