+++ packages/database-editor/src/components/ImportWizard/TableMultiSelect.tsx
@@ -2,9 +2,9 @@ import type { DatabaseTable } from '@axonivy/database-editor-protocol';
 import { Button, DropdownMenu, DropdownMenuCheckboxItem, DropdownMenuContent, DropdownMenuTrigger, IvyIcon } from '@axonivy/ui-components';
 import { IvyIcons } from '@axonivy/ui-icons';
 import { useTranslation } from 'react-i18next';
-import './TableDropdown.css';
+import './TableMultiSelect.css';
 
-export const TableDropdown = ({
+export const TableMultiSelect = ({
   tables,
   selection: selectedTables,
   updateSelection
@@ -17,14 +17,14 @@ export const TableDropdown = ({
   return (
     <DropdownMenu>
       <DropdownMenuTrigger asChild>
-        <Button className='table-dropdown-trigger'>
+        <Button className='table-select-trigger'>
           <span className='table-trigger-content'>
             {selectedTables.length ? selectedTables.map(t => t.name).join(', ') : t('import.selectTables')}
           </span>
           <IvyIcon icon={IvyIcons.Chevron} rotate={90} />
         </Button>
       </DropdownMenuTrigger>
-      <DropdownMenuContent className='table-dropdown-content'>
+      <DropdownMenuContent className='table-select-content'>
         {tables.map((table, i) => (
           <DropdownMenuCheckboxItem
             className='table-selection-item'
