+++ packages/database-editor/src/components/ImportWizard/pages/TableSelectionPage.tsx
@@ -1,13 +1,13 @@
-import type { DatabaseEditorContext, DatabaseEditorDBContext, DatabaseTable } from '@axonivy/database-editor-protocol';
-import { Button, Flex, Label } from '@axonivy/ui-components';
+import type { DatabaseEditorContext, DatabaseTable } from '@axonivy/database-editor-protocol';
+import { BasicField, Button, Flex } from '@axonivy/ui-components';
 import { IvyIcons } from '@axonivy/ui-icons';
 import { useQuery } from '@tanstack/react-query';
-import { useEffect, useMemo, useState } from 'react';
+import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useClient } from '../../../protocol/ClientContextProvider';
 import { genQueryKey } from '../../../query/query-client';
 import { notImplemented } from '../ImportWizard';
-import { TableDropdown } from '../TableDropdown';
+import { TableMultiSelect } from '../TableMultiSelect';
 
 export type SelectTablesPageProps = {
   context: DatabaseEditorContext;
@@ -19,23 +19,16 @@ export type SelectTablesPageProps = {
 export const SelectTablesPage = ({ context, selectedDatabase, updateSelection, selectedTables }: SelectTablesPageProps) => {
   const { t } = useTranslation();
   const client = useClient();
-  const [tableContext, setTableContext] = useState<DatabaseEditorDBContext>({
-    app: context.app,
-    pmv: context.pmv,
-    file: context.file,
-    databaseName: selectedDatabase
-  });
 
-  useEffect(() => {
-    setTableContext(() => {
-      return {
-        app: context.app,
-        pmv: context.pmv,
-        file: context.file,
-        databaseName: selectedDatabase
-      };
-    });
-  }, [selectedDatabase, context]);
+  const tableContext = useMemo(
+    () => ({
+      app: context.app,
+      pmv: context.pmv,
+      file: context.file,
+      databaseName: selectedDatabase
+    }),
+    [context, selectedDatabase]
+  );
 
   const tableQuery = useQuery({
     queryKey: useMemo(() => genQueryKey('datbaseInfo', tableContext), [tableContext]),
@@ -48,13 +41,14 @@ export const SelectTablesPage = ({ context, selectedDatabase, updateSelection, s
 
   return (
     <Flex className='import-page table-selection-page' direction='column'>
-      <div className='import-grid'>
-        <Label className='import-label'>{t('import.selectMainTables')}</Label>
-        <TableDropdown tables={tableQuery.data?.tables ?? []} updateSelection={updateSelection} selection={selectedTables} />
-        <Button onClick={notImplemented} variant='outline' icon={IvyIcons.Plus}>
-          {t('import.add')}
-        </Button>
-      </div>
+      <BasicField label={t('import.selectMainTables')}>
+        <Flex gap={2}>
+          <TableMultiSelect tables={tableQuery.data?.tables ?? []} updateSelection={updateSelection} selection={selectedTables} />
+          <Button onClick={notImplemented} variant='outline' icon={IvyIcons.Plus}>
+            {t('import.add')}
+          </Button>
+        </Flex>
+      </BasicField>
     </Flex>
   );
 };
