+++ packages/database-editor/src/translation/database-editor/de.json
@@ -14,7 +14,6 @@
     "importWizard": "Import Wizard",
     "ivySystemDatabase": "IvySystemDatabase",
     "next": "Zum nächsten Schritt",
-    "noData": "Keine Daten",
     "repository": "DAO/Repo Classes",
     "selectDatabase": "Datenbank auswählen",
     "selectMainTables": "Tabellen auswählen",
