+++ packages/database-editor/src/translation/database-editor/en.json
@@ -14,7 +14,6 @@
     "importWizard": "Import Wizard",
     "ivySystemDatabase": "IvySystemDatabase",
     "next": "Go to next step",
-    "noData": "No data",
     "repository": "DAO/Repo Classes",
     "selectDatabase": "Select a database",
     "selectMainTables": "Select main tables",
