+++ playwright/tests/integration/mock/importWizard/import-wizard.spec.ts
@@ -2,27 +2,27 @@ import { expect, test } from '@playwright/test';
 import { DatabaseEditor } from '../../../pageobjects/DatabaseEditor';
 import type { ImportDialog } from '../../../pageobjects/ImportDialog';
 
-let editor: DatabaseEditor;
-let importDialog: ImportDialog;
-
-test.beforeEach(async ({ page }) => {
-  editor = await DatabaseEditor.openMock(page);
-  importDialog = editor.importDialog;
-});
-
 test.describe('import wizard', () => {
+  let editor: DatabaseEditor;
+  let importDialog: ImportDialog;
+
+  test.beforeEach(async ({ page }) => {
+    editor = await DatabaseEditor.openMock(page);
+    importDialog = editor.importDialog;
+  });
+
   test('opening behaviour', async () => {
-    await importDialog.trigger.expectName('Import Wizard');
+    await expect(importDialog.trigger).toHaveText('Import Wizard');
     await importDialog.open();
     await expect(importDialog.locator.first()).toBeVisible();
   });
 
   test('navigation buttons', async () => {
     await importDialog.open();
-    await importDialog.next.expectName('Go to next step');
-    await importDialog.next.expectDisabled();
-    await importDialog.back.expectName('Back');
-    await importDialog.back.expectDisabled();
+    await expect(importDialog.next).toHaveText('Go to next step');
+    await expect(importDialog.next).toBeDisabled();
+    await expect(importDialog.back).toHaveText('Back');
+    await expect(importDialog.back).toBeDisabled();
   });
 
   test('timeline', async () => {
