+++ playwright/tests/pageobjects/DataSourcePage.ts
@@ -5,12 +5,10 @@ export class DataSourcePage {
   readonly locator: Locator;
   readonly typeSelection: Locator;
   readonly databaseSelect: Select;
-  readonly contentGrid: Locator;
 
   constructor(page: Page, parent: Locator) {
     this.locator = parent.locator('.data-source-page');
-    this.contentGrid = this.locator.locator('.import-grid');
     this.typeSelection = this.locator.locator('.source-group');
-    this.databaseSelect = new Select(page, this.contentGrid, { nth: 0 });
+    this.databaseSelect = new Select(page, this.locator, { nth: 0 });
   }
 }
