+++ playwright/tests/pageobjects/ImportDialog.ts
@@ -1,15 +1,14 @@
-import { type Locator, type Page } from '@playwright/test';
-import { Button } from './Button';
+import { expect, type Locator, type Page } from '@playwright/test';
 import { CreationPage } from './CreationPage';
 import { DataSourcePage } from './DataSourcePage';
 import { TableSelectionPage } from './TableSelectionPage';
 import { Timeline } from './Timeline';
 
 export class ImportDialog {
-  readonly trigger: Button;
-  readonly next: Button;
-  readonly create: Button;
-  readonly back: Button;
+  readonly trigger: Locator;
+  readonly next: Locator;
+  readonly create: Locator;
+  readonly back: Locator;
   readonly locator: Locator;
   readonly timeline: Timeline;
   readonly dataSourcePage: DataSourcePage;
@@ -17,11 +16,11 @@ export class ImportDialog {
   readonly creationPage: CreationPage;
 
   constructor(page: Page, parent: Locator) {
-    this.trigger = new Button(parent, { name: 'Import Wizard' });
     this.locator = parent.getByRole('dialog');
-    this.next = new Button(this.locator, { name: 'Go to next step' });
-    this.create = new Button(this.locator, { name: 'Create' });
-    this.back = new Button(this.locator, { name: 'Back' });
+    this.trigger = parent.getByRole('button', { name: 'Import Wizard' });
+    this.next = this.locator.getByRole('button', { name: 'Go to next step' });
+    this.create = this.locator.getByRole('button', { name: 'Create' });
+    this.back = this.locator.getByRole('button', { name: 'Back' });
     this.timeline = new Timeline(this.locator);
     this.dataSourcePage = new DataSourcePage(page, this.locator);
     this.tableSelectionPage = new TableSelectionPage(page, this.locator);
@@ -30,5 +29,6 @@ export class ImportDialog {
 
   async open() {
     await this.trigger.click();
+    await expect(this.locator).toBeVisible();
   }
 }
