+++ playwright/tests/pageobjects/TableSelectionPage.ts
@@ -4,11 +4,9 @@ import { MultiSelect } from './MultiSelect';
 export class TableSelectionPage {
   readonly locator: Locator;
   readonly tableSelect: MultiSelect;
-  readonly contentGrid: Locator;
 
   constructor(page: Page, parent: Locator) {
     this.locator = parent.locator('.table-selection-page');
-    this.contentGrid = this.locator.locator('.import-grid');
-    this.tableSelect = new MultiSelect(page, this.contentGrid);
+    this.tableSelect = new MultiSelect(page, this.locator);
   }
 }
